Extracted processing command to Alchemist.Helpers.ProcessCommands
It allows us to run process commands from different server
implementation.

Also improved Alchemist.Server.IO to be responsible of processing
commands when a command is entered in STDIN
