+++ Makefile
@@ -8,16 +8,13 @@ STAT_COLOR=\033[2;33m
 
 all: test
 
-test:	test_server
+test:
 	${MAKE} test_helpers
 	${MAKE} test_api
 
-test_server:
-	@ echo "\n$(INFO_COLOR)Run server tests: $(NO_COLOR)\n"
-	$(ELIXIR) test/server_test.exs
-
 test_helpers:
 	@ echo "\n$(INFO_COLOR)Run helper tests: $(NO_COLOR)\n"
+	$(ELIXIR) test/helpers/process_commands_test.exs
 	$(ELIXIR) test/helpers/module_info_test.exs
 	$(ELIXIR) test/helpers/complete_test.exs
