+++ lib/server/io.exs
@@ -1,16 +1,27 @@
+Code.require_file "../helpers/process_commands.exs", __DIR__
+
 defmodule Alchemist.Server.IO do
 
   @moduledoc false
 
-  def gets do
+  use GenServer
+
+  alias Alchemist.Helpers.ProcessCommands
+
+  def start(env) do
+    GenServer.start_link(__MODULE__,  env, [])
+  end
+
+  def init(env) do
+    {:ok, env, 0}
+  end
+
+  def handle_info(:timeout, env) do
+    ProcessCommands.process(read_line, env, Process.group_leader)
+    {:noreply, env, 0}
+  end
+
+  def read_line do
     IO.gets("") |> String.rstrip()
-  end
-
-  def puts(line) do
-    IO.puts line
-  end
-
-  def inspect(item) do
-    IO.inspect item
   end
 end
